Separate common components logic into a function
